local vion adjusted
added a global array that specifies if wich player turns can be seen.
can be switched off if player is controlled remotly or is AI. switch is
maual for now for ndebug purposes.

isObservable - new atribute used to check if assets are graphically
visable

isVisable - legacy attribut that no longer affects graphics but may be
used in controls for acting player.

next on the docket
- make sure workes in game continue